create desi_run_night script
