+++ bin/desi_run_prod
@@ -2,8 +2,6 @@
 # coding: utf-8
 
 import argparse
-import socket
-import sys
 
 from desispec.scripts.submit_prod import submit_production
 
@@ -32,3 +30,4 @@ if __name__ == '__main__':
     args = parse_args()
 
     submit_production(**args.__dict__)
+
